Consolidate secrets.yaml read/merge/write into helpers/secrets_state
